DataFactoryMonitor/requirements.txt
adal==1.2.1
asn1crypto==0.24.0
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
azure-common==1.1.22
azure-mgmt-datafactory==0.7.0
azure-mgmt-resource==3.0.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.4.1
cryptography==2.7
Flask==1.0.3
Flask-WTF==0.14.2
idna==2.8
importlib-metadata==0.18
isodate==0.6.0
isort==4.3.20
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.0.0
msrest==0.6.7
msrestazure==0.6.1
oauthlib==3.0.1
packaging==19.0
pluggy==0.12.0
py==1.8.0
pycparser==2.19
PyJWT==1.7.1
pylint==2.3.1
pyparsing==2.4.0
pytest==4.6.3
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
typed-ast==1.4.0
urllib3==1.25.3
wcwidth==0.1.7
Werkzeug==0.15.4
wrapt==1.11.1
WTForms==2.2.1
zipp==0.5.1
